requirements.txt
pandas==1.1.4
requests==2.22.0
beautifulsoup4==4.9.3
dataclasses==0.8
PyGithub==1.54
